#1401. [Patterns] Map patterns tests updated. Check primitive == for keys
