+++ LanguageFeatures/Patterns/map_A06_t01.dart
@@ -24,28 +24,25 @@
 /// very unlikely that any map pattern containing identical keys (and no rest
 /// element) will ever match. Duplicate keys are most likely a typo in the code.
 ///
-/// If any two keys in the map both have primitive == methods, then it is a
-/// compile-time error if they are equal according to their == operator. In
-/// cases where keys have types whose equality can be checked at compile time,
-/// we report errors if there are redundant keys. But we don't require the keys
-/// to have primitive equality for flexibility. In map patterns where the keys
-/// don't have primitive equality, it is possible to have redundant keys and the
-/// compiler won't detect it.
+/// Any two record keys which both have primitive == are equal. Since records
+/// don't have defined identity, we can't use the previous rule to detect
+/// identical records. But records do support an equality test known at compile
+/// time if all of their fields do, so we use that.
 ///
 /// There is more than one ... element in the map pattern.
 ///
 /// The ... element is not the last element in the map pattern.
 ///
-/// @description Check that it is a compile-time error if the ... element is not
-/// the last element in the map pattern.
+/// @description Check that it is a compile-time error if there is more than one
+/// ... element in the map pattern.
 /// @author sgrekhov22@gmail.com
 
 // SharedOptions=--enable-experiment=patterns
 
 String test1(Map map) {
   return switch (map) {
-    case <int, int>{1: 1, ..., 2: 2} => "";
-//                        ^^^
+    case <int, int>{1: 1, ..., ...} => "";
+//                             ^^^
 // [analyzer] unspecified
 // [cfe] unspecified
     default => "default";
@@ -54,29 +51,29 @@ String test1(Map map) {
 
 void test2(Map map) {
   switch (map) {
-    case {3: 4, ..., 5: 6}:
-//              ^^^
+    case {3: 4, ..., ...}:
+//                   ^^^
 // [analyzer] unspecified
 // [cfe] unspecified
   }
 }
 
 String test3(Map map) {
-  if (map case {1: _, 2: _, ..., 3: _}) {
-//                          ^^^
+  if (map case {1: _, 2: _, ..., ...}) {
+//                               ^^^
 // [analyzer] unspecified
 // [cfe] unspecified
   }
 }
 
 main() {
-  var {1: a, 2: b, ..., 4: 4} = {1: 1, 2: 2, 3: 3, 4: 4};
-//                 ^^^
+  var {1: a, 2: b, ..., ...} = {1: 1, 2: 2, 3: 3, 4: 4};
+//                      ^^^
 // [analyzer] unspecified
 // [cfe] unspecified
 
-  final {..., 4: 4} = {1: 1, 2: 2, 3: 3, 4: 4};
-//       ^^^
+  final {..., ...} = {1: 1, 2: 2, 3: 3, 4: 4};
+//            ^^^
 // [analyzer] unspecified
 // [cfe] unspecified
