+++ LanguageFeatures/Patterns/map_A07_t01.dart
@@ -6,102 +6,78 @@
 /// mapPattern        ::= typeArguments? '{' mapPatternEntries? '}'
 /// mapPatternEntries ::= mapPatternEntry ( ',' mapPatternEntry )* ','?
 /// mapPatternEntry   ::= expression ':' pattern | '...'
-/// ...
-/// Like lists, map patterns can also have a rest pattern. However, there's no
-/// well-defined notion of a map "minus" some set of matched entries. Thus, the
-/// rest pattern doesn't allow a subpattern to capture the "remaining" entries.
 ///
-/// Also, there is no ordering to entries in a map, so we only allow the ... to
-/// appear as the last element. Appearing anywhere else would send a confusing,
-/// meaningless signal.
+/// A map pattern matches values that implement Map and accesses values by key
+/// from it.
 ///
-/// In practice, this means that the only purpose of ... in a map pattern is to
-/// allow matching a map that contains extra entries, while ignoring those
-/// entries. By default, a map pattern only matches if the map's length is
-/// exactly the same as the number of entry subpatterns. Adding a ... element
-/// allows the map pattern to match if the map's length is at least the number
-/// of (non-rest) entry subpatterns (and all of those subpatterns match).
+/// It is a compile-time error if:
 ///
-/// @description Check that if rest pattern is present then map, that has more
-/// elements than number of subpatterns in a map pattern, may match
+/// typeArguments is present and there are more or fewer than two type arguments
+///
+/// Any of the entry key expressions are not constant expressions.
+///
+/// If any two keys in the map are identical. Map patterns that don't have a
+/// rest element only match if the length of the map is equal to the number of
+/// map entries. If a map pattern has multiple identical key entries, they will
+/// increase the required length for the pattern to match but in all but the
+/// most perverse Map implementations will represent the same key. Thus, it's
+/// very unlikely that any map pattern containing identical keys (and no rest
+/// element) will ever match. Duplicate keys are most likely a typo in the code.
+///
+/// Any two record keys which both have primitive == are equal. Since records
+/// don't have defined identity, we can't use the previous rule to detect
+/// identical records. But records do support an equality test known at compile
+/// time if all of their fields do, so we use that.
+///
+/// There is more than one ... element in the map pattern.
+///
+/// The ... element is not the last element in the map pattern.
+///
+/// @description Check that it is a compile-time error if the ... element is not
+/// the last element in the map pattern.
 /// @author sgrekhov22@gmail.com
 
 // SharedOptions=--enable-experiment=patterns
 
-import "../../Utils/expect.dart";
-
 String test1(Map map) {
   return switch (map) {
-    case {1: 1, 2: > 0, ...} => "case-1";
-    case {1: 2, 2: <= 0, ...} => "case-2";
-    case {1: 3, 2: var a, ...} => "case-3";
+    case <int, int>{1: 1, ..., 2: 2} => "";
+//                        ^^^
+// [analyzer] unspecified
+// [cfe] unspecified
     default => "default";
   };
 }
 
-String test2(Map map) {
+void test2(Map map) {
   switch (map) {
-    case {1: 1, 2: > 0, ...}:
-      return "case-1";
-    case {1: 2, 2: <= 0, ...}:
-      return "case-2";
-    case {1: 3, 2: var a, ...}:
-      return "case-3";
-    default:
-      return "default";
+    case {3: 4, ..., 5: 6}:
+//              ^^^
+// [analyzer] unspecified
+// [cfe] unspecified
   }
 }
 
 String test3(Map map) {
-  if (map case {1: 1, 2: > 0, ...}) {
-    return "case-1";
+  if (map case {1: _, 2: _, ..., 3: _}) {
+//                          ^^^
+// [analyzer] unspecified
+// [cfe] unspecified
   }
-  if (map case {1: 2, 2: <= 0, ...}) {
-    return "case-2";
-  }
-  if (map case {1: 3, 2: var a, ...}) {
-    return "case-3";
-  }
-  return "default";
 }
 
 main() {
-  var {1: a, 2: b, ...} = {1: 1, 2: 2, 3: 3};
-  Expect.equals(1, a);
-  Expect.equals(2, b);
+  var {1: a, 2: b, ..., 4: 4} = {1: 1, 2: 2, 3: 3, 4: 4};
+//                 ^^^
+// [analyzer] unspecified
+// [cfe] unspecified
 
-  final {1: 2, 2: int c, ...} = {1: 2, 2: 2, 3: 3};
-  Expect.equals(3, c);
+  final {..., 4: 4} = {1: 1, 2: 2, 3: 3, 4: 4};
+//       ^^^
+// [analyzer] unspecified
+// [cfe] unspecified
 
-  var {1: 3, 2: _, ...} = {1: 3, 2: 2, 3: 3};
-
-  Expect.equals("case-1", test1({1: 1, 2: 1}));
-  Expect.equals("case-1", test1({1: 1, 2: 2, 3: 3}));
-  Expect.equals("default", test1({1: 1}));
-  Expect.equals("case-2", test1({1: 2, 2: 0}));
-  Expect.equals("case-2", test1({1: 2, 2: -1, 3: 3}));
-  Expect.equals("default", test1({2: 0}));
-  Expect.equals("case-3", test1({1: 3, 2: 0}));
-  Expect.equals("case-3", test1({1: 3, 2: -1, 3: 3}));
-  Expect.equals("default", test1({1: 3}));
-
-  Expect.equals("case-1", test2({1: 1, 2: 1}));
-  Expect.equals("case-1", test2({1: 1, 2: 2, 3: 3}));
-  Expect.equals("default", test2({1: 1}));
-  Expect.equals("case-2", test2({1: 2, 2: 0}));
-  Expect.equals("case-2", test2({1: 2, 2: -1, 3: 3}));
-  Expect.equals("default", test2({2: 0}));
-  Expect.equals("case-3", test2({1: 3, 2: 0}));
-  Expect.equals("case-3", test2({1: 3, 2: -1, 3: 3}));
-  Expect.equals("default", test2({1: 3}));
-
-  Expect.equals("case-1", test3({1: 1, 2: 1}));
-  Expect.equals("case-1", test3({1: 1, 2: 2, 3: 3}));
-  Expect.equals("default", test3({1: 1}));
-  Expect.equals("case-2", test3({1: 2, 2: 0}));
-  Expect.equals("case-2", test3({1: 2, 2: -1, 3: 3}));
-  Expect.equals("default", test3({2: 0}));
-  Expect.equals("case-3", test3({1: 3, 2: 0}));
-  Expect.equals("case-3", test3({1: 3, 2: -1, 3: 3}));
-  Expect.equals("default", test3({1: 3}));
+  test1({});
+  test2({});
+  test3({});
 }
